Move runners to distinct dir
