+++ ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/select_positive.py
@@ -19,7 +19,7 @@ from ydb.library.yql.providers.generic.connector.tests.utils.schema import (
     makeOptionalYdbTypeFromYdbType,
 )
 
-from ydb.library.yql.providers.generic.connector.tests.test_cases.select_positive_common import TestCase
+from ydb.library.yql.providers.generic.connector.tests.common_test_cases.select_positive_common import TestCase
 
 
 class Factory:
