+++ ydb/library/yql/providers/generic/connector/tests/datasource/clickhouse/ya.make
@@ -53,6 +53,7 @@ PEERDIR(
     ydb/library/yql/providers/generic/connector/tests/common_test_cases
     ydb/library/yql/providers/generic/connector/tests/utils
     ydb/library/yql/providers/generic/connector/tests/utils/clients
+    ydb/library/yql/providers/generic/connector/tests/utils/run
     ydb/library/yql/providers/generic/connector/tests/utils/scenario
 )
