+++ ydb/library/yql/providers/generic/connector/tests/datasource/postgresql/conftest.py
@@ -3,13 +3,9 @@ import pathlib
 
 import pytest
 
-import yatest.common as yat
 
 from ydb.library.yql.providers.generic.connector.api.common.data_source_pb2 import EDataSourceKind
 from ydb.library.yql.providers.generic.connector.tests.utils.settings import Settings
-from ydb.library.yql.providers.generic.connector.tests.utils.dqrun import DqRunner
-from ydb.library.yql.providers.generic.connector.tests.utils.kqprun import KqpRunner
-from ydb.library.yql.providers.generic.connector.tests.utils.runner import Runner
 from ydb.library.yql.providers.generic.connector.tests.utils.clients.postgresql import Client
 
 
@@ -24,11 +20,3 @@ def settings() -> Settings:
 @pytest.fixture
 def postgresql_client(settings) -> Client:
     return Client(settings.postgresql)
-
-
-def configure_runner(runner, settings) -> Runner:
-    if runner is DqRunner:
-        return DqRunner(dqrun_path=yat.build_path("ydb/library/yql/tools/dqrun/dqrun"), settings=settings)
-    elif runner is KqpRunner:
-        return KqpRunner(kqprun_path=yat.build_path("ydb/tests/tools/kqprun/kqprun"), settings=settings)
-    return None
