+++ ydb/library/yql/providers/generic/connector/tests/datasource/postgresql/test.py
@@ -2,13 +2,11 @@ import pytest
 
 from ydb.library.yql.providers.generic.connector.api.common.data_source_pb2 import EDataSourceKind
 from ydb.library.yql.providers.generic.connector.tests.utils.settings import Settings
-from ydb.library.yql.providers.generic.connector.tests.utils.runner import Runner
-import ydb.library.yql.providers.generic.connector.tests.utils.dqrun as dqrun
-import ydb.library.yql.providers.generic.connector.tests.utils.kqprun as kqprun
+from ydb.library.yql.providers.generic.connector.tests.utils.run.runners import runner_types, configure_runner
 from ydb.library.yql.providers.generic.connector.tests.utils.clients.postgresql import Client
 import ydb.library.yql.providers.generic.connector.tests.utils.scenario.postgresql as scenario
 
-from conftest import configure_runner, docker_compose_dir
+from conftest import docker_compose_dir
 from collection import Collection
 
 import ydb.library.yql.providers.generic.connector.tests.common_test_cases.select_missing_database as select_missing_database
@@ -21,22 +19,19 @@ tc_collection = Collection(
     Settings.from_env(docker_compose_dir=docker_compose_dir, data_source_kinds=[EDataSourceKind.POSTGRESQL])
 )
 
-runners = (dqrun.DqRunner, kqprun.KqpRunner)
-runners_ids = ("dqrun", "kqprun")
 
-
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
+@pytest.mark.parametrize("runner_type", runner_types)
 @pytest.mark.parametrize("test_case", tc_collection.get('select_positive'), ids=tc_collection.ids('select_positive'))
 @pytest.mark.usefixtures("settings")
 @pytest.mark.usefixtures("postgresql_client")
 def test_select_positive(
     request: pytest.FixtureRequest,
     settings: Settings,
-    runner_type: Runner,
+    runner_type: str,
     postgresql_client: Client,
     test_case: select_positive_common.TestCase,
 ):
-    runner = configure_runner(runner=runner_type, settings=settings)
+    runner = configure_runner(runner_type=runner_type, settings=settings)
     scenario.select_positive(
         settings=settings,
         runner=runner,
@@ -46,7 +41,7 @@ def test_select_positive(
     )
 
 
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
+@pytest.mark.parametrize("runner_type", runner_types)
 @pytest.mark.parametrize(
     "test_case", tc_collection.get('select_missing_database'), ids=tc_collection.ids('select_missing_database')
 )
@@ -55,11 +50,11 @@ def test_select_positive(
 def test_select_missing_database(
     request: pytest.FixtureRequest,
     settings: Settings,
-    runner_type: Runner,
+    runner_type: str,
     postgresql_client: Client,
     test_case: select_missing_database.TestCase,
 ):
-    runner = configure_runner(runner=runner_type, settings=settings)
+    runner = configure_runner(runner_type=runner_type, settings=settings)
     scenario.select_missing_table(
         settings=settings,
         runner=runner,
@@ -69,7 +64,7 @@ def test_select_missing_database(
     )
 
 
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
+@pytest.mark.parametrize("runner_type", runner_types)
 @pytest.mark.parametrize(
     "test_case", tc_collection.get('select_missing_table'), ids=tc_collection.ids('select_missing_table')
 )
@@ -78,11 +73,11 @@ def test_select_missing_database(
 def test_select_missing_table(
     request: pytest.FixtureRequest,
     settings: Settings,
-    runner_type: Runner,
+    runner_type: str,
     postgresql_client: Client,
     test_case: select_missing_table.TestCase,
 ):
-    runner = configure_runner(runner=runner_type, settings=settings)
+    runner = configure_runner(runner_type=runner_type, settings=settings)
     scenario.select_missing_table(
         test_name=request.node.name,
         settings=settings,
@@ -92,18 +87,18 @@ def test_select_missing_table(
     )
 
 
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
+@pytest.mark.parametrize("runner_type", runner_types)
 @pytest.mark.parametrize("test_case", tc_collection.get('select_datetime'), ids=tc_collection.ids('select_datetime'))
 @pytest.mark.usefixtures("settings")
 @pytest.mark.usefixtures("postgresql_client")
 def test_select_datetime(
     request: pytest.FixtureRequest,
     settings: Settings,
-    runner_type: Runner,
+    runner_type: str,
     postgresql_client: Client,
     test_case: select_positive_common.TestCase,
 ):
-    runner = configure_runner(runner=runner_type, settings=settings)
+    runner = configure_runner(runner_type=runner_type, settings=settings)
     scenario.select_positive(
         settings=settings,
         runner=runner,
@@ -113,7 +108,7 @@ def test_select_datetime(
     )
 
 
-@pytest.mark.parametrize("runner_type", runners, ids=runners_ids)
+@pytest.mark.parametrize("runner_type", runner_types)
 @pytest.mark.parametrize(
     "test_case",
     tc_collection.get('select_positive_with_schema'),
@@ -124,11 +119,11 @@ def test_select_datetime(
 def test_select_positive_with_schema(
     request: pytest.FixtureRequest,
     settings: Settings,
-    runner_type: Runner,
+    runner_type: str,
     postgresql_client: Client,
     test_case: select_positive_common.TestCase,
 ):
-    runner = configure_runner(runner=runner_type, settings=settings)
+    runner = configure_runner(runner_type=runner_type, settings=settings)
     scenario.select_pg_schema(
         settings=settings,
         runner=runner,
