+++ ydb/library/yql/providers/generic/connector/tests/datasource/postgresql/ya.make
@@ -52,7 +52,7 @@ PEERDIR(
     ydb/library/yql/providers/generic/connector/api/common
     ydb/library/yql/providers/generic/connector/tests/common_test_cases
     ydb/library/yql/providers/generic/connector/tests/utils
-    ydb/library/yql/providers/generic/connector/tests/utils/types
+    ydb/library/yql/providers/generic/connector/tests/utils/run
     ydb/library/yql/providers/generic/connector/tests/utils/clients
     ydb/library/yql/providers/generic/connector/tests/utils/scenario
 )
