+++ ydb/library/yql/providers/generic/connector/tests/join/conftest.py
@@ -4,8 +4,6 @@ import pathlib
 
 import pytest
 
-import yatest.common as yat
-
 from ydb.library.yql.providers.generic.connector.api.common.data_source_pb2 import EDataSourceKind
 from ydb.library.yql.providers.generic.connector.tests.utils.clients.clickhouse import (
     make_client as make_clickhouse_client,
@@ -13,9 +11,6 @@ from ydb.library.yql.providers.generic.connector.tests.utils.clients.clickhouse 
 )
 from ydb.library.yql.providers.generic.connector.tests.utils.clients.postgresql import Client as PostgreSQLClient
 from ydb.library.yql.providers.generic.connector.tests.utils.settings import Settings
-from ydb.library.yql.providers.generic.connector.tests.utils.dqrun import DqRunner
-from ydb.library.yql.providers.generic.connector.tests.utils.kqprun import KqpRunner
-from ydb.library.yql.providers.generic.connector.tests.utils.runner import Runner
 
 
 docker_compose_dir: Final = pathlib.Path("ydb/library/yql/providers/generic/connector/tests/join")
@@ -41,11 +36,3 @@ def clients(settings):
         ClickHouse=make_clickhouse_client(settings=settings.clickhouse),
         PostgreSQL=PostgreSQLClient(settings=settings.postgresql),
     )
-
-
-def configure_runner(runner, settings) -> Runner:
-    if runner is DqRunner:
-        return DqRunner(dqrun_path=yat.build_path("ydb/library/yql/tools/dqrun/dqrun"), settings=settings)
-    elif runner is KqpRunner:
-        return KqpRunner(kqprun_path=yat.build_path("ydb/tests/tools/kqprun/kqprun"), settings=settings)
-    return None
