+++ ydb/library/yql/providers/generic/connector/tests/join/ya.make
@@ -53,8 +53,8 @@ PEERDIR(
     ydb/library/yql/providers/generic/connector/tests/common_test_cases
     ydb/library/yql/providers/generic/connector/tests/utils
     ydb/library/yql/providers/generic/connector/tests/utils/clients
+    ydb/library/yql/providers/generic/connector/tests/utils/run
     ydb/library/yql/providers/generic/connector/tests/utils/scenario
-    ydb/library/yql/providers/generic/connector/tests/utils/types
 )
 
 DEPENDS(
