+++ ydb/library/yql/providers/generic/connector/tests/utils/run/parent.py
@@ -1,20 +1,8 @@
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from pathlib import Path
-from typing import List, Optional
 
 from ydb.library.yql.providers.generic.connector.tests.utils.settings import Settings, GenericSettings
-from ydb.library.yql.providers.generic.connector.tests.utils.schema import Schema, YsonList
-
-
-@dataclass
-class Result:
-    data_out: Optional[YsonList]
-    data_out_with_types: Optional[List]
-    schema: Optional[Schema]
-    stdout: str
-    stderr: str
-    returncode: int
+from ydb.library.yql.providers.generic.connector.tests.utils.run.result import Result
 
 
 class Runner(ABC):
