+++ ydb/library/yql/providers/generic/connector/tests/utils/types/ya.make
@@ -3,6 +3,7 @@ PY3_LIBRARY()
 PY_SRCS(
     clickhouse.py
     postgresql.py
+    ydb.py
 )
 
 END()
