+++ ydb/library/yql/providers/generic/connector/tests/utils/ya.make
@@ -5,19 +5,15 @@ PY_SRCS(
     comparator.py
     data_source_kind.py
     database.py
-    dqrun.py
     docker_compose.py
     generate.py
-    kqprun.py
     log.py
-    runner.py
     schema.py
     settings.py
     sql.py
 )
 
 PEERDIR(
-    contrib/python/Jinja2
     contrib/python/PyYAML
     ydb/library/yql/providers/generic/connector/api/common
     ydb/library/yql/providers/generic/connector/tests/utils/types
@@ -28,5 +24,8 @@ PEERDIR(
 END()
 
 RECURSE_FOR_TESTS(
+    clients
+    run
+    scenario
     types
 )
